Removed preloaded version of Wikipedia page.
App v1.2 included a preloaded version of Wikipedia page to show on
devices not connected to the internet. Unfortunately, this required two
permissions for an app which previously required none, and also tripled
the size of the apk to be downloaded from the Play store.
Hence, I deleted both the permissions and preloaded version of the
Wikipedia page. Henceforth, clicking the Help button on the ActionBar
simply opens the Wikipedia link in the default browser.